refactor: restructure the resource to a different file
